requirements.txt
Django==5.0.2
django-phonenumber-field==7.3.0
phonenumbers==8.13.30
